ROFO-126 feat: 유저 리포트 기록 조회 api 도메인 변경
