+++ data/src/main/java/com/weit2nd/data/source/spot/FoodSpotDataSource.kt
@@ -1,6 +1,5 @@
 package com.weit2nd.data.source.spot
 
-import com.weit2nd.data.model.spot.FoodSpotHistoriesDTO
 import com.weit2nd.data.service.SpotService
 import okhttp3.MultipartBody
 import javax.inject.Inject
@@ -15,18 +14,6 @@ class FoodSpotDataSource @Inject constructor(
         service.reportFoodSpot(
             reportRequest = reportRequest,
             reportPhotos = reportPhotos,
-        )
-    }
-
-    suspend fun getFoodSpotHistories(
-        userId: Long,
-        count: Int,
-        lastItemId: Long?,
-    ): FoodSpotHistoriesDTO {
-        return service.getFoodSpotHistories(
-            userId = userId,
-            size = count,
-            lastId = lastItemId,
         )
     }
