+++ data/src/main/java/com/weit2nd/data/source/user/UserDataSource.kt
@@ -1,5 +1,6 @@
 package com.weit2nd.data.source.user
 
+import com.weit2nd.data.model.spot.FoodSpotHistoriesDTO
 import com.weit2nd.data.model.user.UserDTO
 import com.weit2nd.data.service.UserService
 import javax.inject.Inject
@@ -9,5 +10,17 @@ class UserDataSource @Inject constructor(
 ) {
     suspend fun getMyUserInfo(): UserDTO {
         return service.getMyUserInfo()
+    }
+
+    suspend fun getFoodSpotHistories(
+        userId: Long,
+        count: Int,
+        lastItemId: Long?,
+    ): FoodSpotHistoriesDTO {
+        return service.getFoodSpotHistories(
+            userId = userId,
+            size = count,
+            lastId = lastItemId,
+        )
     }
 }
